♻ 状態管理をjotaiに変更

+++ src/desktop/original-view/hooks/use-chat-history.tsx
@@ -1,106 +1,103 @@
-import { useRecoilCallback, useRecoilValue } from 'recoil';
-import {
-  apiErrorMessageState,
-  chatHistoriesState,
-  inputFilesState,
-  inputTextState,
-  isHistoryDrawerOpenState,
-  pendingRequestsCountState,
-  pluginConfigState,
-  selectedAssistantIndexState,
-  selectedHistoryIdState,
-  selectedHistoryState,
-} from '../states/states';
+import { pluginCommonConfigAtom } from '@/desktop/public-state';
+import { getBase64EncodedImage } from '@/lib/image';
+import { ChatMessage } from '@/lib/static';
 import { deleteAllRecordsByQuery, isGuestSpace, isMobile } from '@konomi-app/kintone-utilities';
 import { produce } from 'immer';
-import { useSnackbar } from 'notistack';
-import { createNewChatHistory, getChatTitle } from '../action';
+import { useAtomValue } from 'jotai';
+import { useAtomCallback } from 'jotai/utils';
 import { nanoid } from 'nanoid';
-import { ChatMessage } from '@/lib/static';
+import { useSnackbar } from 'notistack';
 import OpenAI from 'openai';
-import { getBase64EncodedImage } from '@/lib/image';
+import { useCallback } from 'react';
+import { useRecoilCallback } from 'recoil';
+import { createNewChatHistory, getChatTitle } from '../action';
+import {
+  apiErrorMessageAtom,
+  chatHistoriesAtom,
+  inputFilesAtom,
+  inputTextAtom,
+  isHistoryDrawerOpenAtom,
+  pendingRequestCountAtom,
+  selectedHistoryAtom,
+  selectedHistoryIdAtom,
+  selectedPluginConditionAtom,
+} from '../states/states';
 
 export const useChatHistory = () => {
   const { enqueueSnackbar } = useSnackbar();
-  const histories = useRecoilValue(chatHistoriesState);
+  const histories = useAtomValue(chatHistoriesAtom);
 
-  const setSelectedHistoryId = useRecoilCallback(
-    ({ reset, set }) =>
-      (historyId: string) => {
-        set(selectedHistoryIdState, historyId);
-        reset(apiErrorMessageState);
-        if (isMobile()) {
-          reset(isHistoryDrawerOpenState);
-        }
-      },
-    []
+  const setSelectedHistoryId = useAtomCallback(
+    useCallback((get, set, historyId: string) => {
+      set(selectedHistoryIdAtom, historyId);
+      set(apiErrorMessageAtom, null);
+      if (isMobile()) {
+        set(isHistoryDrawerOpenAtom, false);
+      }
+    }, [])
   );
 
-  const pushMessage = useRecoilCallback(
-    ({ set, snapshot }) =>
-      async (message: ChatMessage) => {
-        const selectedHistory = await snapshot.getPromise(selectedHistoryState);
-        const config = await snapshot.getPromise(pluginConfigState);
-        const assistantIndex = await snapshot.getPromise(selectedAssistantIndexState);
-        const assistant = config.conditions[assistantIndex];
+  const pushMessage = useAtomCallback(
+    useCallback(async (get, set, message: ChatMessage) => {
+      const selectedHistory = get(selectedHistoryAtom);
+      const selectedCondition = get(selectedPluginConditionAtom);
 
-        const historyId = selectedHistory?.id ?? nanoid();
-        const history =
-          selectedHistory ??
-          createNewChatHistory({
-            id: historyId,
-            aiModel: assistant.aiModel,
-            temperature: assistant.temperature,
-            maxTokens: assistant.maxTokens,
-            iconUrl: assistant.aiIcon,
-            title: getChatTitle(message),
-            messages: [],
-          });
-
-        const updatedHistory = produce(history, (draft) => {
-          if (draft.messages.length === 0 && assistant.systemPrompt) {
-            draft.messages.push({ id: nanoid(), role: 'system', content: assistant.systemPrompt });
-          }
-          draft.messages.push(message);
+      const historyId = selectedHistory?.id ?? nanoid();
+      const history =
+        selectedHistory ??
+        createNewChatHistory({
+          id: historyId,
+          aiModel: selectedCondition.aiModel,
+          temperature: selectedCondition.temperature,
+          maxTokens: selectedCondition.maxTokens,
+          iconUrl: selectedCondition.aiIcon,
+          title: getChatTitle(message),
+          messages: [],
         });
 
-        set(selectedHistoryIdState, historyId);
-        set(selectedHistoryState, updatedHistory);
-      },
-    []
+      const updatedHistory = produce(history, (draft) => {
+        if (draft.messages.length === 0 && selectedCondition.systemPrompt) {
+          draft.messages.push({
+            id: nanoid(),
+            role: 'system',
+            content: selectedCondition.systemPrompt,
+          });
+        }
+        draft.messages.push(message);
+      });
+
+      set(selectedHistoryIdAtom, historyId);
+      set(selectedHistoryAtom, updatedHistory);
+    }, [])
   );
 
-  const pushUserMessage = useRecoilCallback(
-    ({ reset, snapshot }) =>
-      async () => {
-        const content = await snapshot.getPromise(inputTextState);
-        const files = await snapshot.getPromise(inputFilesState);
-        if (files) {
-          const imageUrlList = await Promise.all(
-            files.map(async (file) => getBase64EncodedImage(file))
-          );
-          const imageContents: OpenAI.ChatCompletionContentPartImage[] = imageUrlList.map(
-            (url) => ({
-              type: 'image_url',
-              image_url: {
-                url,
-                detail: 'auto',
-              },
-            })
-          );
+  const pushUserMessage = useAtomCallback(
+    useCallback(async (get, set) => {
+      const content = get(inputTextAtom);
+      const files = get(inputFilesAtom);
+      if (files) {
+        const imageUrlList = await Promise.all(
+          files.map(async (file) => getBase64EncodedImage(file))
+        );
+        const imageContents: OpenAI.ChatCompletionContentPartImage[] = imageUrlList.map((url) => ({
+          type: 'image_url',
+          image_url: {
+            url,
+            detail: 'auto',
+          },
+        }));
 
-          await pushMessage({
-            id: nanoid(),
-            role: 'user',
-            content: [{ type: 'text', text: content.replace(/\n/, '  \n') }, ...imageContents],
-          });
-        } else {
-          await pushMessage({ id: nanoid(), role: 'user', content: content.replace(/\n/, '  \n') });
-        }
-        reset(inputTextState);
-        reset(inputFilesState);
-      },
-    []
+        await pushMessage({
+          id: nanoid(),
+          role: 'user',
+          content: [{ type: 'text', text: content.replace(/\n/, '  \n') }, ...imageContents],
+        });
+      } else {
+        await pushMessage({ id: nanoid(), role: 'user', content: content.replace(/\n/, '  \n') });
+      }
+      set(inputTextAtom, '');
+      set(inputFilesAtom, []);
+    }, [])
   );
 
   const pushAssistantMessage = useRecoilCallback(
@@ -110,40 +107,41 @@ export const useChatHistory = () => {
     []
   );
 
-  const removeSelectedHistory = useRecoilCallback(
-    ({ reset, set, snapshot }) =>
-      async () => {
-        try {
-          set(pendingRequestsCountState, (count) => count + 1);
-          const id = (await snapshot.getPromise(selectedHistoryIdState))!;
-          const { common } = await snapshot.getPromise(pluginConfigState);
-          const { outputAppId, outputKeyFieldCode, outputAppSpaceId } = common;
-
-          const isGuest = await isGuestSpace(outputAppId);
-
-          const query = `${outputKeyFieldCode} = "${id}"`;
-
-          await deleteAllRecordsByQuery({
-            app: outputAppId,
-            query,
-            debug: process.env.NODE_ENV === 'development',
-            guestSpaceId: isGuest ? outputAppSpaceId : undefined,
-          });
-
-          set(chatHistoriesState, (_histories) =>
-            produce(_histories, (draft) => {
-              const index = draft.findIndex((history) => history.id === id);
-              draft.splice(index, 1);
-            })
-          );
-
-          reset(selectedHistoryIdState);
-          enqueueSnackbar('履歴を削除しました', { variant: 'success' });
-        } finally {
-          set(pendingRequestsCountState, (count) => count - 1);
+  const removeSelectedHistory = useAtomCallback(
+    useCallback(async (get, set) => {
+      try {
+        set(pendingRequestCountAtom, (count) => count + 1);
+        const id = get(selectedHistoryIdAtom);
+        if (!id) {
+          return;
         }
-      },
-    []
+        const common = useAtomValue(pluginCommonConfigAtom);
+        const { outputAppId, outputKeyFieldCode, outputAppSpaceId } = common;
+
+        const isGuest = await isGuestSpace(outputAppId);
+
+        const query = `${outputKeyFieldCode} = "${id}"`;
+
+        await deleteAllRecordsByQuery({
+          app: outputAppId,
+          query,
+          debug: process.env.NODE_ENV === 'development',
+          guestSpaceId: isGuest ? outputAppSpaceId : undefined,
+        });
+
+        set(chatHistoriesAtom, (_histories) =>
+          produce(_histories, (draft) => {
+            const index = draft.findIndex((history) => history.id === id);
+            draft.splice(index, 1);
+          })
+        );
+
+        set(selectedHistoryIdAtom, null);
+        enqueueSnackbar('履歴を削除しました', { variant: 'success' });
+      } finally {
+        set(pendingRequestCountAtom, (count) => count - 1);
+      }
+    }, [])
   );
 
   return {
